Integrated SearchIO.BlatIO into id_ranker()
Added a check to id_ranker for the type of record given to it.
Added a function based on the UCSC percent identity algorithm as
programmed by BioPython to get the percent identity of an entire hit
as opposed to on a per-HSP basis.
Added functions to filter BLAT hits based on number of HSPs,
query length coverage, percent identity, and scores.

Signed-off-by: docmanny <juan@vazquez.bio>