[V1.4] Initial BLAT compatibility release
Blat has been integrated from start to finish, and the whole program
runs from start to finish as well.
Because SearchIO is experimental, this release requires the use of the
master branch of Biopython, version 1.70dev0, with a commit of 3d6e111
or newer due to previous issues with reading blatx output in either psl
or pslx format.
Additionally, the non-standard package multiprocess by uqfoundation
(https://github.com/uqfoundation/multiprocess) was used to replace
multiprocessing, since multiprocessing uses the pickle package which
is incompatible with SearchIO due to Biopython issue #1241. When a
release fixing the pickling issue comes out, I will retest the use of
multiprocessing and weigh returning to it to keep non-standard package
usage to a minimum.

Signed-off-by: docmanny <juan@vazquez.bio>