New print() function with built-in indent and markup params; PEP8 stuff
print() has been enhanced with options to indent the print statement
to make the indent statements in the code neater to read.
Also of note is the markup keyword, which will be placed between the
indent and the text. This is for future formatting of output as Markdown
files - on the to-do list is to edit this function to be a bit smarter
about how it combines indents and markups so that things look solid in
their final form.
Also did some PEP8 maintenance from the last several edits.

Signed-off-by: docmanny <juan@vazquez.bio>