Various improvements
- hit_name_only: boolean, allows RecBlast to only save the hit names and
ranges in fasta format
- min_mem: boolean, dumps RBC after writing to disk
- RecBlastContainer now correctly prints in a neat format, concatentates
with other RecBlastContainers via '+' and 'sum()' functions, and has
a built-in 'write' method as opposed to requiring an external function.
- RecBlastMP now has a function for accepting forward BLAST/BLAT output
files as input if one only wants to do a Reverse BLAST.
- RecBlast no longer dumps the entire FASTA-formatted sequence into the
output logs, and instead only prints the first few nucleotides.
- 'Best-Hit' method of RecBlast now correctly annotates file
- perc_ident can now be inputted as a decimal, and will be converted to
an integer internally to satisfy BLAST's weird requirements.
- Various comments and exception handles added.