The long-awaited "Everything Works"* update.
* Terms and Conditions may apply.
- Modified filter functions hit_hsp_span and hit_minscore to be stricter
- Added hit_same_strand filter function that returns hits that have all
HSPs on the same strand; added associated functionality for that too.
- LOTS OF RecBlastContainer CHANGES! Much more friendly to initialize,
can make a full RBC from init, includes query species info;
- Began work on a generate_stats method;
- ADDED A RESULTS_FILTER METHOD TO RBC! Applies a given filter function
to each 'recblast_results' object within the RBC tree; has keyword arg
'summary_statistic' that takes a function, and applies it to a specified
part of the RBC according to the 'recblast_object' kwarg.
- Added a write method to the RBC, as well as individual private _write
functions that it calls, to write the RBC as a variety of types.
- Fixed __add__ method so that it correctly add together RBC objects;
will not overwrite a pre-existing species-query RBC object; likewise
fixed __radd__
- id_search now correctly finds assembly-type and
chromosome-type IDs; also identifies and returns strand info.
- Adds SeqFeatures to the SeqRecords specifying that these are
duplicates; the range of the sequence; and the reverse annotation.
- Correctly orients strand and sequence range numbering
- Added 'perc_span' kwarg to relevant functions to allow recblast_MP()
to control parameter.
- Minor error handling improvements
- Greatly expanded documentation with examples
- Options to declare output folder name up-front.