RBC improvements, including RBC_dump and RBC_load
RBC now has an __iter__ method that returns tuples of (species, query,
record).
RBC_dump and RBC_load now operate by pickling each record piece-by-piece
and appending a cumulative checksum to the end of the file. Note that
its up to the user to save said checksum. RBC_load has a "ignore" option
for the checksum that should only be used as a last-ditched solution if
one forgets the checksum themselves.