Big RecSearch commit; running well post paradigm shift
Separated parameters for forward and reverse searches
Renamed RecBlast to RecSearch
Re-added start/stop functionality in the program
Moved some parts of the RecBlast worker from the main function to
class methods
Changed how id_lists are handled; now they are formatted as BED6 files
for consistency.
BIG CHANGE: recblast_run is now the RecSearch object, which simplifies
idiomatically how the program is run. Now, rather than a blocky wall of
parameters, RecSearch is configured step-by-step and is executed as a
separate object.

Signed-off-by: docmanny <juan@vazquez.bio>